Updated to `0.5.0` (#6)
- Removed the `world` fallback mechanism for emission factors.
- Added and updated emission factors.
- Removed deprecated emission factors and the `deprecated` and `replacedBy` properties.

_NB: this version is not retro-compatible (it's not possible to compute values for older versions with this implementation). This has been done to enable reducing code and cleaning up the emission factors list._

### Removed the `world` fallback mechanism for emission factors

In previous versions, the model may use a default value for an emission factor which was not available for the product's country. Some materials had an emission factor associated with the `world`, which would then serve as default value.

In the beginning, this was useful to simplify managing emission factors. Every new product was manually reviewed, so errors could not be introduced. To import a growing number of products that cannot be all manually reviewed, we need to remove this fallback mechanism that may introduce silent errors. For example, if a new (material, country) combination is introduced, the model may continue to work even if the appropriate emission factor is not present in the parameters, due to the presence of the default value.

Emission factors are now attached to a list of country IDs. This list makes it explicit to use a given emission factor for a given country.

### Added and updated emission factors

- Emission factors have been added to support values for new countries and new materials.
- Emission factors have been reviewed to ensure they were up-to-date and as accurate as possible. The reviewed emission factors have been updated to model version `0.5.0`.

### Removed old and deprecated emission factors, and the `deprecated` and `replacedBy` properties

The modelling of emission factors and model parameters has been simplified to make it easier to use and less error-prone.

- Emission factors are attached to a list of countries (instead of "connected entities"),
- `deprecated` and `isReplacedBy` properties have been removed:

  - For auditing purposes, the versioning of the model parameters via Git could be used instead of keeping old and deprecated emission factors.
  - While this could be used to make more dynamic changes in emission factors, until it was implemented in the model, this could only be a source of errors (e.g. by using a deprecated emission factor).

History of emission factors (previous model versions) have been removed too.